lint: Move normalizeDocParamName from parameter-utils to doc-comment
Rationale:
The parameter-utils/ directory contained only a single function
(normalizeDocParamName) that is exclusively used for normalizing GML
documentation parameter names. The name 'parameter-utils' was too generic
for a single-function directory that is tightly coupled to doc-comment
processing.

Original location: src/lint/src/parameter-utils/
New location: src/lint/src/doc-comment/normalize-param-name.ts

The doc-comment directory is the logical home for this function because:
1. The function is specifically for GML doc comment parameter normalization
2. It is used exclusively by feather rules and the normalize-doc-comments rule
3. It shares semantic territory with documented-params.ts which has a related
   function (normalizeDocParamNameForComparison)

Changes:
- Moved parameter-utils/index.ts to doc-comment/normalize-param-name.ts
- Updated imports in lint/src/index.ts
- Updated imports in rules/feather/create-feather-rule.ts
- Updated imports in rules/gml/rules/normalize-doc-comments-rule.ts
- Added export in doc-comment/index.ts
- Removed old parameter-utils directory
