+++ src/lint/src/rules/gml/rules/normalize-doc-comments-rule.ts
@@ -1,7 +1,7 @@
 import { Core } from "@gmloop/core";
 import type { Rule } from "eslint";
 
-import { normalizeDocParamName } from "../../../parameter-utils/index.js";
+import { normalizeDocParamName } from "../../../doc-comment/normalize-param-name.js";
 import { gmlRuleDocCommentServices } from "../gml-rule-services.js";
 import type { GmlRuleDefinition } from "../index.js";
 import {
